requirements.txt
# Core LangChain Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=2.0.0
langgraph>=0.0.20

# Google AI Dependencies
google-generativeai>=0.8.0

# Financial Data & Analysis
yfinance>=0.2.0
pandas>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# News & Media APIs
newsapi-python>=0.2.7
newspaper3k>=0.2.8

# Web Scraping & HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# SEC Filing Access
sec-edgar-downloader>=5.0.0

# Vector Database for RAG
chromadb>=0.4.0

# Web Search Dependencies
duckduckgo-search>=5.0.0
requests>=2.31.0
